+++ docs/generate_diagrams.jl
@@ -53,7 +53,7 @@ end
   a1 <|-left- PolynomialBasis
 
   together {
-    struct UniformPolyBasis {
+    struct CartProdPolyBasis {
       +get_exponents
       +get_orders
     }
@@ -66,7 +66,7 @@ end
     }
   }
 
-  PolynomialBasis <|-- UniformPolyBasis
+  PolynomialBasis <|-- CartProdPolyBasis
   PolynomialBasis <|-- CompWiseTensorPolyBasis
   PolynomialBasis <|-- RaviartThomasPolyBasis
   PolynomialBasis <|-- NedelecPolyBasisOnSimplex
@@ -77,7 +77,7 @@ end
   object "QGrad[<:Polynomial]Basis\nQCurlGrad[<:Polynomial]Basis" as m2
   object "PCurlGrad[<:Polynomial]Basis" as m4
   object "PGradMonomialBasis" as m5
-  UniformPolyBasis <-down- m1
+  CartProdPolyBasis <-down- m1
   CompWiseTensorPolyBasis <-down- m2
   RaviartThomasPolyBasis <-down- m4
   NedelecPolyBasisOnSimplex <-down- m5