+++ src/Polynomials/CartProdPolyBases.jl
@@ -3,15 +3,16 @@
 #################################
 
 """
-    struct UniformPolyBasis{D,V,K,PT} <: PolynomialBasis{D,V,K,PT}
+    struct CartProdPolyBasis{D,V,K,PT} <: PolynomialBasis{D,V,K,PT}
 
-Type representing a uniform basis of (an)isotropic `D`-multivariate `V`-valued
-polynomial space
+"Cartesian product polynomial basis"
+
+Type representing a basis of a (an)isotropic `D`-multivariate `V`-valued
+cartesian product polynomial space
 
 `V`(𝕊, 𝕊, ..., 𝕊)
 
-where 𝕊 is a scalar multivariate polynomial space. So each (independant)
-component of `V` holds the same space (hence the name 'uniform').
+where 𝕊 is a scalar multivariate polynomial space.
 
 The scalar polynomial basis spanning 𝕊 is defined as
 
@@ -22,15 +23,14 @@ evaluated at xᵢ (iᵗʰ comp. of x), and where α = (α₁, α₂, ..., α`D`)
 multi-index in `terms`, a subset of ⟦0,`K`⟧`ᴰ`. `terms` is a field that can be
 passed in a constructor.
 
-The fields of this `struct` are not public.
 This type fully implements the [`Field`](@ref) interface, with up to second
 order derivatives.
 """
-struct UniformPolyBasis{D,V,K,PT} <: PolynomialBasis{D,V,K,PT}
+struct CartProdPolyBasis{D,V,K,PT} <: PolynomialBasis{D,V,K,PT}
   orders::NTuple{D,Int}
   terms::Vector{CartesianIndex{D}}
 
-  function UniformPolyBasis{D}(
+  function CartProdPolyBasis{D}(
     ::Type{PT},
     ::Type{V},
     orders::NTuple{D,Int},
@@ -45,20 +45,20 @@ struct UniformPolyBasis{D,V,K,PT} <: PolynomialBasis{D,V,K,PT}
   end
 end
 
-@inline Base.size(a::UniformPolyBasis{D,V}) where {D,V} = (length(a.terms)*num_indep_components(V),)
+@inline Base.size(a::CartProdPolyBasis{D,V}) where {D,V} = (length(a.terms)*num_indep_components(V),)
 
-function UniformPolyBasis(
+function CartProdPolyBasis(
    ::Type{PT},
    ::Val{D},
    ::Type{V},
    orders::NTuple{D,Int},
    terms::Vector{CartesianIndex{D}}) where {PT<:Polynomial,D,V}
 
-  UniformPolyBasis{D}(PT,V,orders,terms)
+  CartProdPolyBasis{D}(PT,V,orders,terms)
 end
 
 """
-    UniformPolyBasis(::Type{PT}, ::Val{D}, ::Type{V}, orders::Tuple [, filter=_q_filter])
+    CartProdPolyBasis(::Type{PT}, ::Val{D}, ::Type{V}, orders::Tuple [, filter=_q_filter])
 
 This constructor allows to pass a tuple `orders` containing the maximal
 polynomial order to be used in each of the `D` spatial dimensions in order to
@@ -67,35 +67,35 @@ construct a tensorial anisotropic `D`-multivariate space 𝕊.
 If a filter is provided, it is applied on the cartesian product terms
 CartesianIndices(`orders`), with maximum(`orders`) as order argument.
 """
-function UniformPolyBasis(
+function CartProdPolyBasis(
   ::Type{PT}, ::Val{D}, ::Type{V}, orders::NTuple{D,Int}, filter::Function=_q_filter
   ) where {PT,D,V}
 
   terms = _define_terms(filter, orders)
-  UniformPolyBasis{D}(PT,V,orders,terms)
+  CartProdPolyBasis{D}(PT,V,orders,terms)
 end
 
 """
-    UniformPolyBasis(::Type{PT}, ::Type{V}, ::Val{D}, order::Int [, filter=_q_filter])
+    CartProdPolyBasis(::Type{PT}, ::Type{V}, ::Val{D}, order::Int [, filter=_q_filter])
 
-Return a `UniformPolyBasis{D,V,order,PT}` where 𝕊 is defined by the terms
+Return a `CartProdPolyBasis{D,V,order,PT}` where 𝕊 is defined by the terms
 filtered by
 
     term -> filter(term, order).
 
 See the [Filter functions](@ref) section of the documentation for more details.
 """
-function UniformPolyBasis(
+function CartProdPolyBasis(
   ::Type{PT}, VD::Val{D}, ::Type{V}, order::Int, filter::Function=_q_filter) where {PT,D,V}
 
   orders = tfill(order,VD)
-  UniformPolyBasis(PT,Val(D),V,orders,filter)
+  CartProdPolyBasis(PT,Val(D),V,orders,filter)
 end
 
 # API
 
 """
-    get_exponents(b::UniformPolyBasis)
+    get_exponents(b::CartProdPolyBasis)
 
 Get a vector of tuples with the exponents of all the terms in the basis of 𝕊,
 the components scalar space of `b`.
@@ -115,17 +115,17 @@ println(exponents)
 Tuple{Int,Int}[(0, 0), (1, 0), (2, 0), (0, 1), (1, 1), (2, 1), (0, 2), (1, 2), (2, 2)]
 ```
 """
-function get_exponents(b::UniformPolyBasis)
+function get_exponents(b::CartProdPolyBasis)
   indexbase = 1
   [Tuple(t) .- indexbase for t in b.terms]
 end
 
 """
-    get_orders(b::UniformPolyBasis)
+    get_orders(b::CartProdPolyBasis)
 
 Return the D-tuple of polynomial orders in each spatial dimension
 """
-function get_orders(b::UniformPolyBasis)
+function get_orders(b::CartProdPolyBasis)
   b.orders
 end
 
@@ -134,7 +134,7 @@ end
 #################################
 
 function _evaluate_nd!(
-  b::UniformPolyBasis{D,V,K,PT}, x,
+  b::CartProdPolyBasis{D,V,K,PT}, x,
   r::AbstractMatrix{V}, i,
   c::AbstractMatrix{T}) where {D,V,K,PT,T}
 
@@ -157,22 +157,22 @@ function _evaluate_nd!(
       @inbounds s *= c[d,ci[d]]
     end
 
-    k = _uniform_set_value!(r,i,s,k)
+    k = _cartprod_set_value!(r,i,s,k)
   end
 end
 
 """
-    _uniform_set_value!(r::AbstractMatrix{<:Real},i,s,k)
+    _cartprod_set_value!(r::AbstractMatrix{<:Real},i,s,k)
 
 r[i,k] = s; return k+1
 """
-function _uniform_set_value!(r::AbstractMatrix{<:Real},i,s,k)
+function _cartprod_set_value!(r::AbstractMatrix{<:Real},i,s,k)
   @inbounds r[i,k] = s
   k+1
 end
 
 """
-    _uniform_set_value!(r::AbstractMatrix{V},i,s::T,k,l)
+    _cartprod_set_value!(r::AbstractMatrix{V},i,s::T,k,l)
 
 ```
 r[i,k]     = V(s, 0,    ..., 0)
@@ -184,7 +184,7 @@ return k+N
 
 where `N = num_indep_components(V)`.
 """
-function _uniform_set_value!(r::AbstractMatrix{V},i,s::T,k) where {V,T}
+function _cartprod_set_value!(r::AbstractMatrix{V},i,s::T,k) where {V,T}
   ncomp = num_indep_components(V)
   z = zero(T)
   @inbounds for j in 1:ncomp
@@ -195,7 +195,7 @@ function _uniform_set_value!(r::AbstractMatrix{V},i,s::T,k) where {V,T}
 end
 
 function _gradient_nd!(
-  b::UniformPolyBasis{D,V,K,PT}, x,
+  b::CartProdPolyBasis{D,V,K,PT}, x,
   r::AbstractMatrix{G}, i,
   c::AbstractMatrix{T},
   g::AbstractMatrix{T},
@@ -226,12 +226,12 @@ function _gradient_nd!(
       end
     end
 
-    k = _uniform_set_derivative!(r,i,s,k,V)
+    k = _cartprod_set_derivative!(r,i,s,k,V)
   end
 end
 
 """
-    _uniform_set_derivative!(r::AbstractMatrix{G},i,s,k,::Type{<:Real})
+    _cartprod_set_derivative!(r::AbstractMatrix{G},i,s,k,::Type{<:Real})
 
 ```
 r[i,k] = s = (∇bᵏ)(xi); return k+1
@@ -240,7 +240,7 @@ r[i,k] = s = (∇bᵏ)(xi); return k+1
 where bᵏ is the kᵗʰ basis polynomial. Note that `r[i,k]` is a `VectorValue` or
 `TensorValue` and `s` a `MVector` or `MMatrix` respectively, of same size.
 """
-function _uniform_set_derivative!(
+function _cartprod_set_derivative!(
   r::AbstractMatrix{G},i,s,k,::Type{<:Real}) where G
 
   @inbounds r[i,k] = s
@@ -248,7 +248,7 @@ function _uniform_set_derivative!(
 end
 
 """
-    _uniform_set_derivative!(r::AbstractMatrix{G},i,s,k,::Type{V})
+    _cartprod_set_derivative!(r::AbstractMatrix{G},i,s,k,::Type{V})
 
 ```
 z = zero(s)
@@ -262,7 +262,7 @@ return k+n
 Note that `r[i,k]` is a `TensorValue` or `ThirdOrderTensorValue` and `s` a
 `MVector` or `MMatrix`.
 """
-@generated function _uniform_set_derivative!(
+@generated function _cartprod_set_derivative!(
   r::AbstractMatrix{G},i,s,k,::Type{V}) where {G,V}
   # Git blame me for readable non-generated version
 
@@ -297,7 +297,7 @@ end
 # necessary as long as outer(Point, V<:AbstractSymTensorValue)::G does not
 # return a tensor type G that implements the appropriate symmetries of the
 # gradient (and hessian)
-@generated function _uniform_set_derivative!(
+@generated function _cartprod_set_derivative!(
   r::AbstractMatrix{G},i,s,k,::Type{V}) where {G,V<:AbstractSymTensorValue{D}} where D
   # Git blame me for readable non-generated version
 
@@ -336,7 +336,7 @@ end
 end
 
 function _hessian_nd!(
-  b::UniformPolyBasis{D,V,K,PT}, x,
+  b::CartProdPolyBasis{D,V,K,PT}, x,
   r::AbstractMatrix{G}, i,
   c::AbstractMatrix{T},
   g::AbstractMatrix{T},
@@ -373,7 +373,7 @@ function _hessian_nd!(
       end
     end
 
-    k = _uniform_set_derivative!(r,i,s,k,V)
+    k = _cartprod_set_derivative!(r,i,s,k,V)
   end
 end