+++ src/Polynomials/MonomialBases.jl
@@ -8,11 +8,11 @@ struct Monomial <: Polynomial   end
 isHierarchical(::Type{Monomial}) = true
 
 """
-    MonomialBasis{D,V,K} = UniformPolyBasis{D,V,K,Monomial}
+    MonomialBasis{D,V,K} = CartProdPolyBasis{D,V,K,Monomial}
 
-Alias for monomial Multivariate scalar' or `Multivalue`'d basis.
+Alias for cartesian product monomial basis, scalar valued or multi-valued.
 """
-const MonomialBasis{D,V,K} = UniformPolyBasis{D,V,K,Monomial}
+const MonomialBasis{D,V,K} = CartProdPolyBasis{D,V,K,Monomial}
 
 """
     MonomialBasis(::Val{D}, ::Type{V}, order::Int, terms::Vector)
@@ -21,7 +21,7 @@ const MonomialBasis{D,V,K} = UniformPolyBasis{D,V,K,Monomial}
 
 High level constructors of [`MonomialBasis`](@ref).
 """
-MonomialBasis(args...) = UniformPolyBasis(Monomial, args...)
+MonomialBasis(args...) = CartProdPolyBasis(Monomial, args...)
 
 function PGradBasis(::Type{Monomial},::Val{D},::Type{T},order::Int) where {D,T}
   NedelecPolyBasisOnSimplex{D}(Monomial,T,order)
@@ -29,7 +29,7 @@ end
 function PGradBasis(::Type{Monomial},::Val{1},::Type{T},order::Int) where T
   @check T<:Real "T needs to be <:Real since represents the type of the components of the vector value"
   V = VectorValue{1,T}
-  UniformPolyBasis(Monomial, Val(1), V, order+1)
+  CartProdPolyBasis(Monomial, Val(1), V, order+1)
 end
 
 # 1D evaluation implementation