+++ test/PolynomialsTests/PCurlGradBasesTests.jl
@@ -73,7 +73,7 @@ T = Float64
 V = VectorValue{D,T}
 b = PCurlGradBasis(PT,Val(D),T,order)
 
-@test b isa UniformPolyBasis{D,V,order+1,PT}
+@test b isa CartProdPolyBasis{D,V,order+1,PT}
 
 @test_throws AssertionError PCurlGradBasis(PT,Val(D),V,order)